requirements.txt
streamlit==1.35.0
chromadb==0.5.0
sentence-transformers==3.0.0
groq==0.9.0
pdfplumber==0.11.0
requests==2.32.3
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.0
